% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_games.R
\name{retrieve_chesscom_games}
\alias{retrieve_chesscom_games}
\title{Retrieve Chess.com Games}
\usage{
retrieve_chesscom_games(players)
}
\arguments{
\item{players}{One or more Chess.com usernames}
}
\value{
A dataframe where all elements are characters.
}
\description{
This function retrieves rated games for one or more players from chess.com.
}
