% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_games.R
\name{retrieve_lichess_games}
\alias{retrieve_lichess_games}
\title{Retrieve Lichess Games}
\usage{
retrieve_lichess_games(player)
}
\arguments{
\item{player}{A Lichess username}
}
\value{
A dataframe where all elements are characters
}
\description{
This function retrieves the bullet, blitz, rapid and classical rated games for a single player
}
