% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_games.R
\name{retrieve_lichess_games}
\alias{retrieve_lichess_games}
\title{Retrieve Lichess Games}
\usage{
retrieve_lichess_games(players)
}
\arguments{
\item{player}{One or more Lichess usernames}
}
\value{
A dataframe where all elements are characters
}
\description{
This function retrieves rated games for one or more players from Lichess.
}
